Holy moly this was a lot (READ THE DESCRIPTION!!!!1111111!!11!11)
Alright, here we go. So I'm really close to finishing the pivot mechanism in this version. What I've done is that I've added parameters to the enum constantsin the PivotTarget class, created a control mode class with setters and getters, and added a lot to periodic so now every 20ms it checks the mode that the robot is in, and which control mode state it is in, and will go to a certain position based on the mode that it is in (I still need to add placeholder values). Also, I've added runCharacterization in another class (which i'm still trying to figure out what it does and if it actually has a use). That's about it though.